% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{fabric_disconnect}
\alias{fabric_disconnect}
\title{Disconnect from Fabric}
\usage{
fabric_disconnect(con)
}
\arguments{
\item{con}{A DBI connection object}
}
\value{
Invisible TRUE
}
\description{
Closes the database connection
}
\examples{
\dontrun{
fabric_disconnect(con)
}
}
