% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fabric_list_databases}
\alias{fabric_list_databases}
\title{List All Available Databases in Fabric}
\usage{
fabric_list_databases(con)
}
\arguments{
\item{con}{A DBI connection object (connected to master database)}
}
\value{
A data frame with database information
}
\description{
Retrieves a list of all databases the user has access to in the Fabric endpoint
}
\examples{
\dontrun{
# Connect to endpoint (no database specified)
con <- fabric_connect_ad("your-endpoint.datawarehouse.fabric.microsoft.com")

# List all databases
databases <- fabric_list_databases(con)
print(databases)

fabric_disconnect(con)
}
}
