% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{fabric_list_tables}
\alias{fabric_list_tables}
\title{List Tables in Fabric Database}
\usage{
fabric_list_tables(con)
}
\arguments{
\item{con}{A DBI connection object}
}
\value{
A character vector of table names
}
\description{
Retrieves a list of all tables in the connected Fabric database
}
\examples{
\dontrun{
tables <- fabric_list_tables(con)
print(tables)
}
}
