% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_stored_password}
\alias{get_stored_password}
\title{Get Stored Password}
\usage{
get_stored_password(email)
}
\arguments{
\item{email}{Email to verify against stored credentials}
}
\value{
Decrypted password or NULL
}
\description{
Retrieves and decrypts stored password
}
\keyword{internal}
