% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{load_fabric_credentials}
\alias{load_fabric_credentials}
\title{Load Stored Fabric Credentials}
\usage{
load_fabric_credentials()
}
\value{
List with email and password, or NULL if not found
}
\description{
Loads stored credentials from environment variables
}
\keyword{internal}
