% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{store_fabric_credentials}
\alias{store_fabric_credentials}
\title{Store Fabric Credentials Securely}
\usage{
store_fabric_credentials(email, password)
}
\arguments{
\item{email}{User email}

\item{password}{User password}
}
\description{
Stores email and password using environment variables with encryption
}
\keyword{internal}
